requirements.txt
aniso8601==9.0.1
apturl==0.5.2
asgiref==3.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.3
duplicity==0.8.21
fasteners==0.14.1
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.38.0
future==0.18.2
httplib2==0.20.2
idna==3.3
image==1.5.33
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
matplotlib==3.6.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
py-mon==1.1.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
systemd-python==234
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
watchdog==2.1.9
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
zipp==1.0.0
